scylla-connectivity-geoip/requirements.txt
scylla-driver
geoip2
python-geohash